test: 小切手使用の入金失敗時ロールバック原子性テストを追加(MySQL)
受取人の名前を varchar(16) 超にして入金 SaveChanges を失敗させ、
Used 更新も巻き戻って小切手が未使用のまま残ることを検証する
(旧 UseAsync の補償欠落バグが単一tx化で解消されたことを担保)。
ファイル/クラス名を ChequeUseMySqlTests へ改名し並行+原子性を集約。
